+++ src/cmd/script.rs
@@ -187,6 +187,14 @@ pub fn new() -> Command {
                 .num_args(1)
                 .value_parser(value_parser!(String)),
         )
+        .arg(
+            Arg::new("global_incdir")
+                .long("global-incdir")
+                .help("Add a global include directory")
+                .num_args(1)
+                .action(ArgAction::Append)
+                .value_parser(value_parser!(String)),
+        )
 }
 
 fn get_package_strings<I>(packages: I) -> IndexSet<String>
@@ -439,9 +447,25 @@ fn emit_template(
                 .iter()
                 .map(|(k, &v)| (k.to_string(), v.map(String::from))),
         );
-        all_incdirs.append(&mut src.clone().get_incdirs());
+        all_incdirs.append(
+            &mut src
+                .clone()
+                .get_incdirs()
+                .iter()
+                .map(|p| p.to_path_buf())
+                .collect(),
+        );
         all_files.append(&mut src.files.clone());
     }
+
+    let global_incdirs: Vec<PathBuf> =
+        if let Some(incdirs) = &matches.get_many::<String>("global_incdir") {
+            incdirs.clone().map(Into::into).collect()
+        } else {
+            [].to_vec()
+        };
+    all_incdirs.append(&mut global_incdirs.clone().into_iter().collect());
+
     let all_defines = if !matches.get_flag("only-includes") && !matches.get_flag("only-sources") {
         all_defines
     } else {
@@ -451,7 +475,7 @@ fn emit_template(
 
     let all_incdirs: IndexSet<PathBuf> =
         if !matches.get_flag("only-defines") && !matches.get_flag("only-sources") {
-            all_incdirs.into_iter().map(|p| p.to_path_buf()).collect()
+            all_incdirs.into_iter().collect()
         } else {
             IndexSet::new()
         };
@@ -493,12 +517,22 @@ fn emit_template(
                         );
                         local_defines
                     },
-                    incdirs: src
-                        .clone()
-                        .get_incdirs()
-                        .iter()
-                        .map(|p| p.to_path_buf())
-                        .collect(),
+                    incdirs: {
+                        let mut incdir_index_list: Vec<PathBuf> = src
+                            .clone()
+                            .get_incdirs()
+                            .iter()
+                            .map(|p| p.to_path_buf())
+                            .collect();
+
+                        incdir_index_list.append(&mut global_incdirs.clone().into_iter().collect());
+
+                        incdir_index_list
+                            .into_iter()
+                            .collect::<IndexSet<PathBuf>>()
+                            .into_iter()
+                            .collect()
+                    },
                     files: files
                         .iter()
                         .map(|f| match f {